# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: kmcheck.com VIN Lookup API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/vin-catalog:
    post:
      operationId: adminCreateVinCatalog
      tags: [admin]
      summary: Manually create a new VIN catalog entry (does not call any provider)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VinDataPatch"
                - type: object
                  required: [vin]
                  properties:
                    vin:
                      type: string
                      minLength: 17
                      maxLength: 17
                    providerName:
                      type: string
                      nullable: true
      responses:
        "201":
          description: Created catalog entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinCatalogItem"
        "400":
          description: Invalid VIN
        "409":
          description: VIN already exists in catalog

    get:
      operationId: adminGetVinCatalog
      tags: [admin]
      summary: List all VIN catalog entries (one per unique VIN)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: vin
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated VIN catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVinCatalogPage"
  /admin/vin-catalog/import:
    post:
      operationId: adminImportVinCatalog
      tags: [admin]
      summary: Import VIN catalog entries from a CSV file (smart merge)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import result summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinImportResult"
        "400":
          description: Invalid file or no valid VINs
  /admin/vin-catalog/import-json:
    post:
      operationId: adminImportVinCatalogJson
      tags: [admin]
      summary: Import VIN catalog entries from a full-data JSON export (smart merge, preserves all fields)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import result summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinImportResult"
        "400":
          description: Invalid JSON file or no valid VINs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VinCatalogItem:
      type: object
      required: [id, vin, data, importedAt, updatedAt]
      properties:
        id:
          type: integer
        vin:
          type: string
        providerName:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
        importedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminVinCatalogPage:
      type: object
      required: [items, total, page, limit, providerStats]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VinCatalogItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        providerStats:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
                nullable: true
              cnt:
                type: integer
...
    VinImportResult:
      type: object
      required: [inserted, updated, skipped, total, conflicts]
      properties:
        inserted:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        total:
          type: integer
        conflicts:
          type: array
          items:
            type: object
            properties:
              vin:
                type: string
              existing:
                type: string
              incoming:
                type: string
...
    VinDataPatch:
      type: object
      description: Partial VIN data fields to merge into the existing catalog record
      properties:
        make:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        trim:
          type: string
          nullable: true
        engine:
          type: string
          nullable: true
        transmission:
          type: string
          nullable: true
        fuelType:
          type: string
          nullable: true
        bodyType:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        odometer:
          type: number
          nullable: true
        ownerCount:
          type: integer
          nullable: true
        accidentCount:
          type: integer
          nullable: true
        hp:
          type: number
          nullable: true
        cylinders:
          type: integer
          nullable: true
        titleStatus:
          type: string
          nullable: true
        isSalvage:
          type: boolean
        isStolen:
          type: boolean
        country:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            type: string
        accidents:
          nullable: true
        insuranceClaims:
          nullable: true
        mileageHistory:
          nullable: true
        ownerHistory:
          nullable: true
        auctionHistory:
          nullable: true
        registryHistory:
          nullable: true
        marketData:
          nullable: true
...